RangeActionSensitivityProvider now extends LoadflowProvider instead of the other way round
